LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-start-end-expected.txt


Harness Error (TIMEOUT), message = null

PASS Sanity check for testValue length; if this fails, variou absolute offsets in the test below need to be adjusted to be less than testValue.length
PASS Initial .value set on input-appended should set selectionStart to end of value
PASS Initial .value set on input-not-appended should set selectionStart to end of value
PASS Initial .value set on input-appended-prefocused should set selectionStart to end of value
PASS Initial .value set on input-not-appended-prefocused should set selectionStart to end of value
PASS Initial .value set on textarea-appended should set selectionStart to end of value
PASS Initial .value set on textarea-not-appended should set selectionStart to end of value
PASS Initial .value set on textarea-appended-prefocused should set selectionStart to end of value
PASS Initial .value set on textarea-not-appended-prefocused should set selectionStart to end of value
PASS Initial .value set on input-appended should set selectionEnd to end of value
PASS Initial .value set on input-not-appended should set selectionEnd to end of value
PASS Initial .value set on input-appended-prefocused should set selectionEnd to end of value
PASS Initial .value set on input-not-appended-prefocused should set selectionEnd to end of value
PASS Initial .value set on textarea-appended should set selectionEnd to end of value
PASS Initial .value set on textarea-not-appended should set selectionEnd to end of value
PASS Initial .value set on textarea-appended-prefocused should set selectionEnd to end of value
PASS Initial .value set on textarea-not-appended-prefocused should set selectionEnd to end of value
PASS onselect should fire when selectionStart is changed
NOTRUN onselect should fire when selectionStart is changed on input-appended
NOTRUN onselect should fire when selectionStart is changed on input-not-appended
NOTRUN onselect should fire when selectionStart is changed on input-appended-prefocused
NOTRUN onselect should fire when selectionStart is changed on input-not-appended-prefocused
NOTRUN onselect should fire when selectionStart is changed on textarea-appended
NOTRUN onselect should fire when selectionStart is changed on textarea-not-appended
NOTRUN onselect should fire when selectionStart is changed on textarea-appended-prefocused
NOTRUN onselect should fire when selectionStart is changed on textarea-not-appended-prefocused
NOTRUN onselect should fire when selectionStart is changed on
PASS onselect should fire when selectionEnd is changed
NOTRUN onselect should fire when selectionEnd is changed on input-appended
NOTRUN onselect should fire when selectionEnd is changed on input-not-appended
NOTRUN onselect should fire when selectionEnd is changed on input-appended-prefocused
NOTRUN onselect should fire when selectionEnd is changed on input-not-appended-prefocused
NOTRUN onselect should fire when selectionEnd is changed on textarea-appended
NOTRUN onselect should fire when selectionEnd is changed on textarea-not-appended
NOTRUN onselect should fire when selectionEnd is changed on textarea-appended-prefocused
NOTRUN onselect should fire when selectionEnd is changed on textarea-not-appended-prefocused
NOTRUN onselect should fire when selectionEnd is changed on
PASS Setting selectionStart to a value larger than selectionEnd should increase selectionEnd
PASS Setting selectionEnd to a value smaller than selectionStart should decrease selectionStart
FAIL selectionStart edge-case values assert_equals: selectionStart setter on input-appended should convert -1 to 2^32-1 expected 10 but got 0
FAIL selectionEnd edge-case values assert_equals: selectionEnd setter on input-appended should convert -1 to 2^32-1 expected 10 but got 0
PASS selectionStart should be clamped by the current value length
PASS selectionEnd should be clamped by the current value length
PASS setSelectionRange should be clamped by the current value length
PASS selectionStart and selectionEnd should remain the same when selectionDirection is changed
